Version 2.1 (#8)
* feat: MeasurementService: working Read, Update, Post using MongoDB database.

* Implemented PatientService (#2)

* Create patient integration

* Added MongoDBContext to PatientService
* Added dependencies to Domain
* Added DI of services to Program.cs
* Populated PatientController

* implemented get by id + delete

* Included getPatientWithMeasurements

* Update PatientController.cs

* Update Measurement.cs

* Squashed commit of the following:

commit 617e34be1cf974e6458f12a9dda51f25b35e5ae6
Author: Ádám Lőrincz <90683098+Ladam0203@users.noreply.github.com>
Date:   Mon Apr 22 13:03:38 2024 +0200

    Uncommented compose

commit 769c36621d2ed2f61b5d25c14a65fff49ee28d26
Author: Ádám Lőrincz <90683098+Ladam0203@users.noreply.github.com>
Date:   Mon Apr 22 13:02:33 2024 +0200

    PatientUI can send measurements

commit 607571ecef460bdd6c49bd29c58df8bd6282001c
Author: Ádám Lőrincz <90683098+Ladam0203@users.noreply.github.com>
Date:   Mon Apr 22 12:37:15 2024 +0200

    Added back messed up merge

commit b182e3f729a46d3aaa768138242eadc419232ccd
Merge: 845229c e37876b
Author: Ádám Lőrincz <90683098+Ladam0203@users.noreply.github.com>
Date:   Mon Apr 22 12:33:16 2024 +0200

    Merge branch 'PatientUI-vue' of https://github.com/TawfikAzza/MinSygehus into PatientUI-vue

commit 845229c1edd89efeeead1edca1d5292c4ca2b3d2
Author: Ádám Lőrincz <90683098+Ladam0203@users.noreply.github.com>
Date:   Thu Apr 18 13:18:50 2024 +0200

    Vue UI added, CORS error calling service

commit e37876b53d1153d3e6fade25f3d4ae22aacd3da3
Author: Ádám Lőrincz <90683098+Ladam0203@users.noreply.github.com>
Date:   Thu Apr 18 13:18:50 2024 +0200

    Vue UI added, CORS error calling service

* fix!: adapted MeasurementServiceController to get measurements by ssn + Minor fixes to patientservice to have it running locally as well (for testing). (#3)

* Refactor/measurement service architecture (#4)

* Stabilized measurement service

* Fixed context + usage of environments

* format

* Fixed post measurement + put measurement

* Implemented GetMeasurementBySsn

* DoctorUI - CreatePatient - test happy case

* Endpoints are called from DoctorUI

* Fully worked through functionality

* Sexy extra vite stuff

* Update docker-image.yml

* Version to 2.0

* Feature/monitor logging (#6)

* feat: MeasurementService: working Read, Update, Post using MongoDB database.

* Implemented PatientService (#2)

* Create patient integration

* Added MongoDBContext to PatientService
* Added dependencies to Domain
* Added DI of services to Program.cs
* Populated PatientController

* implemented get by id + delete

* Included getPatientWithMeasurements

* Update PatientController.cs

* Update Measurement.cs

* Squashed commit of the following:

commit 617e34be1cf974e6458f12a9dda51f25b35e5ae6
Author: Ádám Lőrincz <90683098+Ladam0203@users.noreply.github.com>
Date:   Mon Apr 22 13:03:38 2024 +0200

    Uncommented compose

commit 769c36621d2ed2f61b5d25c14a65fff49ee28d26
Author: Ádám Lőrincz <90683098+Ladam0203@users.noreply.github.com>
Date:   Mon Apr 22 13:02:33 2024 +0200

    PatientUI can send measurements

commit 607571ecef460bdd6c49bd29c58df8bd6282001c
Author: Ádám Lőrincz <90683098+Ladam0203@users.noreply.github.com>
Date:   Mon Apr 22 12:37:15 2024 +0200

    Added back messed up merge

commit b182e3f729a46d3aaa768138242eadc419232ccd
Merge: 845229c e37876b
Author: Ádám Lőrincz <90683098+Ladam0203@users.noreply.github.com>
Date:   Mon Apr 22 12:33:16 2024 +0200

    Merge branch 'PatientUI-vue' of https://github.com/TawfikAzza/MinSygehus into PatientUI-vue

commit 845229c1edd89efeeead1edca1d5292c4ca2b3d2
Author: Ádám Lőrincz <90683098+Ladam0203@users.noreply.github.com>
Date:   Thu Apr 18 13:18:50 2024 +0200

    Vue UI added, CORS error calling service

commit e37876b53d1153d3e6fade25f3d4ae22aacd3da3
Author: Ádám Lőrincz <90683098+Ladam0203@users.noreply.github.com>
Date:   Thu Apr 18 13:18:50 2024 +0200

    Vue UI added, CORS error calling service

* fix!: adapted MeasurementServiceController to get measurements by ssn + Minor fixes to patientservice to have it running locally as well (for testing). (#3)

* Refactor/measurement service architecture (#4)

* Stabilized measurement service

* Fixed context + usage of environments

* format

* Fixed post measurement + put measurement

* Implemented GetMeasurementBySsn

* DoctorUI - CreatePatient - test happy case

* Endpoints are called from DoctorUI

* Fully worked through functionality

* Sexy extra vite stuff

* Beginning of trace and logging

* Added tracing to patientS. and measurementS. getBySsn

* Update docker-image.yml

* Version to 2.0

* FeatureHub start

* feature hub

* doctorpos is still false

* Fixed cors

* FeatureHub Solved ! Thanks Adam the man

---------


* Delete Patient and Create Patient FEatureHubbed

* UI disables create and delete

* Added extra log, UI message correction

* 2.1 version

---------

+++ PatientService/Controllers/PatientController.cs
@@ -26,21 +26,14 @@ public class PatientController : ControllerBase {
     [HttpPost]
     public async Task<ActionResult<Patient>> CreatePatient(Patient patient)
     {
-
-        Console.WriteLine("");
-        Console.WriteLine("");
-        Console.WriteLine("");
-        Console.WriteLine("");
-        Console.WriteLine("");
-        Console.WriteLine("");
-        Console.WriteLine("Doctor post: " + _clientContext["doctorPost"].IsEnabled);
-        Console.WriteLine("");
-        Console.WriteLine("");
-        Console.WriteLine("");
-        Console.WriteLine("");
-        Console.WriteLine("");
-
         using var activity = _tracer.StartActiveSpan("CreatePatient");
+        
+        if (!_clientContext["createpatient"].IsEnabled)
+        {
+            Monitoring.Monitoring.Log.Warning("Create patient is disabled");
+            return StatusCode(418, "Create patient is disabled");
+        }
+        
         var result = await _patientManager.Create(patient);
         
         if (result is null) {
@@ -82,8 +75,14 @@ public class PatientController : ControllerBase {
 
     [HttpDelete]
     public async Task<IActionResult> DeletePatient(string ssn) {
+        if (!_clientContext["deletepatient"].IsEnabled)
+        {
+            Monitoring.Monitoring.Log.Warning("Delete patient is disabled");
+            return StatusCode(418, "Delete patient is disabled");
+        }
+        
         var result = await _patientManager.DeleteBySsn(ssn);
-
+        
         if (!result) {
             Monitoring.Monitoring.Log.Error("Couldn't delete a patient");
             return BadRequest($"Couldn't delete the patient with ssn: {ssn}");
